gcoap: Move from gnrc directory up to net directory
