+++ sys/Makefile
@@ -99,6 +99,9 @@ endif
 ifneq (,$(filter sema,$(USEMODULE)))
     DIRS += sema
 endif
+ifneq (,$(filter gcoap,$(USEMODULE)))
+    DIRS += net/application_layer/coap
+endif
 
 DIRS += $(dir $(wildcard $(addsuffix /Makefile, ${USEMODULE})))
