+++ sys/net/gnrc/Makefile
@@ -136,8 +136,5 @@ endif
 ifneq (,$(filter gnrc_tftp,$(USEMODULE)))
     DIRS += application_layer/tftp
 endif
-ifneq (,$(filter gcoap,$(USEMODULE)))
-    DIRS += application_layer/coap
-endif
 
 include $(RIOTBASE)/Makefile.base
